requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

aioredis==2.0.1 ; python_version >= '3.6'
annotated-types==0.7.0 ; python_version >= '3.8'
anyio==4.6.2.post1 ; python_version >= '3.9'
async-timeout==4.0.3 ; python_version >= '3.7'
bcrypt==4.2.0
certifi==2024.8.30 ; python_version >= '3.6'
click==8.1.7 ; python_version >= '3.7'
distlib==0.3.9
dnspython==2.7.0 ; python_version >= '3.9'
ecdsa==0.19.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==2.2.0
exceptiongroup==1.2.2 ; python_version < '3.11'
fastapi==0.115.3 ; python_version >= '3.8'
filelock==3.16.1 ; python_version >= '3.8'
h11==0.14.0 ; python_version >= '3.7'
idna==3.10 ; python_version >= '3.6'
motor==3.6.0 ; python_version >= '3.8'
packaging==24.1 ; python_version >= '3.8'
passlib[bcrypt]==1.7.4
pbr==6.1.0 ; python_version >= '2.6'
pipenv-to-requirements==0.9.0
pipenv==2024.2.0 ; python_version >= '3.8'
platformdirs==4.3.6 ; python_version >= '3.8'
pyasn1==0.6.1 ; python_version >= '3.8'
pydantic-core==2.23.4 ; python_version >= '3.8'
pydantic-settings==2.6.0 ; python_version >= '3.8'
pydantic[email]==2.9.2 ; python_version >= '3.8'
pyjwt==2.9.0 ; python_version >= '3.8'
pymongo==4.9.2 ; python_version >= '3.8'
python-dotenv==1.0.1 ; python_version >= '3.8'
python-jose==3.3.0
python-multipart==0.0.12 ; python_version >= '3.8'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
setuptools==75.2.0 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1 ; python_version >= '3.7'
starlette==0.41.0 ; python_version >= '3.8'
typing-extensions==4.12.2 ; python_version >= '3.8'
uvicorn==0.32.0 ; python_version >= '3.8'
virtualenv==20.27.0 ; python_version >= '3.8'
